pyproject.toml
[tool.poetry]
name = "dpds-sidecar"
version = "0.1.0"
description = ""
authors = ["Mark Stella <mark.p.stella@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rpyc = "^6.0.0"
typer-slim = {extras = ["standard"], version = "^0.12.3"}

[tool.poetry.scripts]
server = "dpds_sidecar.cache:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
